+++ src/EventListener/CreateInvoiceOnOrderPaymentPaidListener.php
@@ -14,10 +14,10 @@ declare(strict_types=1);
 namespace Sylius\InvoicingPlugin\EventListener;
 
 use Sylius\InvoicingPlugin\Creator\InvoiceCreatorInterface;
-use Sylius\InvoicingPlugin\Event\OrderPlaced;
+use Sylius\InvoicingPlugin\Event\OrderPaymentPaid;
 use Sylius\InvoicingPlugin\Exception\InvoiceAlreadyGenerated;
 
-final class CreateInvoiceOnOrderPlacedListener
+final class CreateInvoiceOnOrderPaymentPaidListener
 {
     private InvoiceCreatorInterface $invoiceCreator;
 
@@ -26,7 +26,7 @@ final class CreateInvoiceOnOrderPlacedListener
         $this->invoiceCreator = $invoiceCreator;
     }
 
-    public function __invoke(OrderPlaced $event): void
+    public function __invoke(OrderPaymentPaid $event): void
     {
         try {
             $this->invoiceCreator->__invoke($event->orderNumber(), $event->date());